Cleanup sub-sdk tab/button rendering.
