+++ app/lib/frontend/templates/layout.dart
@@ -234,132 +234,6 @@ String renderSdkTabs({
       'shared/search_tabs', searchTabs.toJson());
 }
 
-class _FilterOption {
-  final String label;
-  final String tag;
-  final String title;
-
-  _FilterOption({
-    @required this.label,
-    @required this.tag,
-    @required this.title,
-  });
-}
-
-String renderSubSdkTabsHtml({
-  @required SearchForm searchForm,
-  bool detectAdvanced = false,
-  bool onlyAdvanced = false,
-}) {
-  final pred = searchForm?.tagsPredicate;
-  if (searchForm?.sdk == SdkTagValue.dart) {
-    return _renderFilterTabs(
-      searchForm: searchForm,
-      options: [
-        if (!onlyAdvanced)
-          _FilterOption(
-            label: 'native',
-            tag: DartSdkTag.runtimeNativeJit,
-            title:
-                'Packages compatible with Dart running on a native platform (JIT/AOT)',
-          ),
-        if (!onlyAdvanced)
-          _FilterOption(
-            label: 'JS',
-            tag: DartSdkTag.runtimeWeb,
-            title: 'Packages compatible with Dart compiled for the web',
-          ),
-      ],
-    );
-  } else if (searchForm?.sdk == SdkTagValue.flutter) {
-    return _renderFilterTabs(
-      searchForm: searchForm,
-      options: [
-        if (!onlyAdvanced)
-          _FilterOption(
-            label: 'Android',
-            tag: FlutterSdkTag.platformAndroid,
-            title: 'Packages compatible with Flutter on the Android platform',
-          ),
-        if (!onlyAdvanced)
-          _FilterOption(
-            label: 'iOS',
-            tag: FlutterSdkTag.platformIos,
-            title: 'Packages compatible with Flutter on the iOS platform',
-          ),
-        if (!onlyAdvanced)
-          _FilterOption(
-            label: 'Web',
-            tag: FlutterSdkTag.platformWeb,
-            title: 'Packages compatible with Flutter on the Web platform',
-          ),
-        // `Linux`, `macOS`, `Windows` platforms are not yet stable, and we want
-        // to display them only when the user has already opted-in to get them
-        // displayed.
-        // TODO: The conditional predicate must be removed once the platform becomes stable.
-        if (onlyAdvanced ||
-            (detectAdvanced &&
-                pred != null &&
-                pred.isRequiredTag(FlutterSdkTag.platformLinux)))
-          _FilterOption(
-            label: 'Linux',
-            tag: FlutterSdkTag.platformLinux,
-            title: 'Packages compatible with Flutter on the Linux platform',
-          ),
-        if (onlyAdvanced ||
-            (detectAdvanced &&
-                pred != null &&
-                pred.isRequiredTag(FlutterSdkTag.platformMacos)))
-          _FilterOption(
-            label: 'macOS',
-            tag: FlutterSdkTag.platformMacos,
-            title: 'Packages compatible with Flutter on the macOS platform',
-          ),
-        if (onlyAdvanced ||
-            (detectAdvanced &&
-                pred != null &&
-                pred.isRequiredTag(FlutterSdkTag.platformWindows)))
-          _FilterOption(
-            label: 'Windows',
-            tag: FlutterSdkTag.platformWindows,
-            title: 'Packages compatible with Flutter on the Windows platform',
-          ),
-      ],
-    );
-  }
-  return null;
-}
-
-String _renderFilterTabs({
-  @required SearchForm searchForm,
-  @required List<_FilterOption> options,
-}) {
-  if (options.isEmpty) return null;
-  final tp = searchForm.tagsPredicate;
-  String searchWithTagsLink(TagsPredicate tagsPredicate) {
-    return searchForm.change(tagsPredicate: tagsPredicate).toSearchLink();
-  }
-
-  final searchTabs = SearchTabs(
-    tabs: options
-        .map((option) => SearchTab(
-              title: option.title,
-              text: option.label,
-              href: htmlAttrEscape.convert(searchWithTagsLink(
-                tp.isRequiredTag(option.tag)
-                    ? tp.withoutTag(option.tag)
-                    : tp.appendPredicate(TagsPredicate(
-                        requiredTags: [option.tag],
-                      )),
-              )),
-              active: tp.isRequiredTag(option.tag),
-            ))
-        .toList(),
-  );
-  return templateCache.renderTemplate(
-      'shared/search_tabs', searchTabs.toJson());
-}
-
 final String _defaultPageDescriptionEscaped = htmlEscape.convert(
     'Pub is the package manager for the Dart programming language, containing reusable '
     'libraries & packages for Flutter, AngularDart, and general Dart programs.');
